fix dto's

+++ src/main/java/org/trackitall/trackitall/supply/dto/SupplyOrderItemRequestDTO.java
@@ -2,12 +2,15 @@
 package org.trackitall.trackitall.supply.dto;
 
 import jakarta.validation.constraints.NotNull;
+import jdk.dynalink.linker.LinkerServices;
 import lombok.Data;
+
+import java.util.List;
 
 @Data
 public class SupplyOrderItemRequestDTO {
     @NotNull(message = "L'ID de la matière première est obligatoire")
-    private Long rawMaterialId;
+    private Integer rawMaterialId;
 
     @NotNull(message = "La quantité est obligatoire")
     private Integer quantity;
